created stabler least squares solver
This least squares solver uses QR factorization to solve the least squares problem. I implemented this because the condition number of transpose(A)*A is roughly the square of the condition number of A, so the normal equations are poorly conditioned relative to the Rx=Q'b problem (since cond(R) = cond(A) and cond(Q) = 1).