inv iter: normalize before and after projecting
Before, the nullspace was giving this solver some trouble because power iteration would pretty much delete all the components in the nullspace direction. As a result, projecting out the existing eigenvectors would leave the new vector *very* tiny, so the backward error was very small in magnitude even if rounding errors caused the eigenvector to be pointing very far off target.