added InverseIterateTime() routine
Now, instead of specifying the precision you want (or not specifying anything at all), you can specify the amount of time you are willing to wait (which is usually what you care about anyway).

Since oscillation-based convergence detection tends to be slow (e.g. it forces an extra ~50 iterations for a 50x50 matrix), it is much slower than it probably needs to be. By giving the user the option to specify a hard timeout, they can decide to sacrifice potential accuracy for time. Usually, this will turn out to be acceptable, and may still yield the same accuracy as oscillation anyway.