refactor inverse iteration code
The old API exposed too much internals of the eigenvalue solver. Specifically, it claimed to be using inverse iteration, which the user should not know (and which was not even entirely accurate). It also required the user to specify a maximum number of iterations, which, as my last commit shows, is very hard to calculate properly.

Now, the API is simpler and more powerful. Specifically, it provides mechanisms to get eigenvectors and eigenvalues asynchronously as they are computed, and to cancel said computations. This could theoretically allow the user to run precision-constrained eigenvalue computation at the same time as maximum-precision computation, selecting whichever values arrive first.