created iterative bisector structure
This might not seem useful now, but the paradigm of having a struct with Step(), Done(), and Root() methods will make it possible to run multiple root-finding algorithms in parallel without requiring a Func to be currency-safe and without weird scheduling/timing issues. In other words, I plan to make another structure like this for other root-finding techniques so that a single function can loop and do one iteration for each technique.