crafted malicious Dekker case
The function 2^-x really screws over the secant method, because each time you go to the right by 1, you go down by a factor of two. In other words, each secant is twice as close to the axis, but always sloped in such a way that the next approximate root is the previous approximation + 1.